+++ Vodovoz/Domain/Operations/GoodsMovementOperation.cs
@@ -71,9 +71,9 @@ namespace Vodovoz.Domain.Operations
 		}
 
 
-		int amount;
+		decimal amount;
 
-		public virtual int Amount {
+		public virtual decimal Amount {
 			get { return amount; }
 			set { SetField (ref amount, value, () => Amount); }
 		}
